+++ src/engine-factory.ts
@@ -10,14 +10,15 @@ import { DevToolEventListener } from './diagnostics/devtool-event-lister';
 import { Diagnostics } from './diagnostics/diagnostics';
 import {
   DEV_TOOLS_KEY,
-  IDiagnosticsInfo,
   TDiagnosticType,
+  TWindowWithDevtools,
 } from './diagnostics/types';
 import {
   ActionRegistryEventbusListener,
   Eventbus,
   IEventbus,
   RequestVideoUriInterceptor,
+  TEventbusRemover,
 } from './eventbus';
 import { LanguageManager } from './language-manager';
 import { TimelineEventNames } from './timeline-event-names';
@@ -34,42 +35,52 @@ import {
 import { prepareValueForSerialization } from './util/prepare-value-for-serialization';
 
 export class EngineFactory implements IEngineFactory {
-  private resizeTimeout: any = -1;
-  private actionsLookup: Record<string, IAction> = {};
-  private importer: ISimpleResourceImporter;
-  private eventbus: IEventbus;
-  private _handleSpaceBound: any;
+  private _actionsLookup: Record<string, IAction> = {};
+  private _eventbus: IEventbus;
+  private _eventRemovers: TEventbusRemover[] = [];
+  private _importer: ISimpleResourceImporter;
+  private _resizeTimeout: any = -1;
 
   constructor(
     importer: ISimpleResourceImporter,
     windowRef: Window,
     options?: IEngineFactoryOptions
   ) {
-    this.importer = importer;
-    this.eventbus = options?.eventbus || new Eventbus();
+    this._importer = importer;
+    this._eventbus = options?.eventbus ?? new Eventbus();
 
     Diagnostics.active = options?.devtools ?? false;
     if (Diagnostics.active) {
-      this._initializeDevtools(this.eventbus);
+      this._initializeDevtools(this._eventbus);
     }
 
-    this.eventbus.on(
-      TimelineEventNames.REQUEST_INSTANCE,
-      this._requestInstanceHandler.bind(this)
+    this._eventRemovers.push(
+      this._eventbus.on(
+        TimelineEventNames.REQUEST_INSTANCE,
+        this._requestInstanceHandler.bind(this)
+      )
     );
-    this.eventbus.on(
-      TimelineEventNames.REQUEST_ACTION,
-      this._requestActionHandler.bind(this)
+    this._eventRemovers.push(
+      this._eventbus.on(
+        TimelineEventNames.REQUEST_ACTION,
+        this._requestActionHandler.bind(this)
+      )
     );
-    this.eventbus.on(
-      TimelineEventNames.REQUEST_FUNCTION,
-      this._requestFunctionHandler.bind(this)
+    this._eventRemovers.push(
+      this._eventbus.on(
+        TimelineEventNames.REQUEST_FUNCTION,
+        this._requestFunctionHandler.bind(this)
+      )
     );
 
-    this._handleSpaceBound = this._handleSpacePress.bind(this);
-    hotkeys('space', this._handleSpaceBound);
+    const handleSpaceBound = this._handleSpacePress.bind(this);
+    hotkeys('space', handleSpaceBound);
+    this._eventRemovers.push(() => hotkeys.unbind('space', handleSpaceBound));
 
-    $(windowRef).resize(this._resizeHandler.bind(this));
+    const resizeBound = this._resizeHandler.bind(this);
+    const jqWin = $(windowRef);
+    jqWin.on('resize', resizeBound);
+    this._eventRemovers.push(() => jqWin.off('resize', resizeBound));
   }
 
   private _handleSpacePress(
@@ -77,20 +88,47 @@ export class EngineFactory implements IEngineFactory {
     _hotkeysEvent: HotkeysEvent
   ): void | boolean {
     keyboardEvent.preventDefault();
-    this.eventbus.broadcast(TimelineEventNames.PLAY_TOGGLE_REQUEST);
+    this._eventbus.broadcast(TimelineEventNames.PLAY_TOGGLE_REQUEST);
     return false;
   }
 
   private _initializeDevtools(eventbus: IEventbus) {
-    const diagnosticInfo = (window as any)[DEV_TOOLS_KEY] as
-      | IDiagnosticsInfo
-      | undefined;
+    const diagnosticInfo = (window as unknown as TWindowWithDevtools)[
+      DEV_TOOLS_KEY
+    ];
+
     if (diagnosticInfo) {
       const { agent } = diagnosticInfo;
-      const eventbusListener = new DevToolEventListener(agent);
-      eventbus.registerEventlistener(eventbusListener);
-      Diagnostics.send = (name: TDiagnosticType, data: any) => {
-        const message = prepareValueForSerialization(data);
+
+      this._eventRemovers.push(
+        eventbus.registerEventlistener(new DevToolEventListener(agent))
+      );
+
+      this._eventRemovers.push(
+        agent.subscribe((message) => {
+          if (message.type === 'playcontrol') {
+            switch (message.data.kind) {
+              case 'play':
+                this._eventbus.broadcast(TimelineEventNames.PLAY_REQUEST);
+                break;
+              case 'stop':
+                this._eventbus.broadcast(TimelineEventNames.STOP_REQUEST);
+                break;
+              case 'pause':
+                this._eventbus.broadcast(TimelineEventNames.PAUSE_REQUEST);
+                break;
+              case 'seek':
+                this._eventbus.broadcast(TimelineEventNames.SEEK_REQUEST, [
+                  message.data.args,
+                ]);
+                break;
+            }
+          }
+        })
+      );
+
+      Diagnostics.send = (name: TDiagnosticType, messageData: any) => {
+        const message = prepareValueForSerialization(messageData);
         try {
           agent.postMessage(name, message);
         } catch (e) {
@@ -99,38 +137,39 @@ export class EngineFactory implements IEngineFactory {
           console.log('message', message);
         }
       };
+
       Diagnostics.send(
         'eligius-diagnostics-factory',
         'Diagnostics initialized'
       );
     } else {
       console.warn(
-        `${DEV_TOOLS_KEY} property not found on window, please install the extension.`
+        `${DEV_TOOLS_KEY} property not found on window, please install the Eligius Devtools extension.`
       );
     }
   }
 
   destroy() {
-    this.eventbus.clear();
-    hotkeys.unbind('space', this._handleSpaceBound);
+    this._eventRemovers.forEach((x) => x());
   }
 
   private _resizeHandler() {
-    if (this.resizeTimeout) {
-      clearTimeout(this.resizeTimeout);
+    if (this._resizeTimeout) {
+      clearTimeout(this._resizeTimeout);
+      this._resizeTimeout = -1;
     }
-    this.resizeTimeout = setTimeout(() => {
-      this.eventbus.broadcast(TimelineEventNames.RESIZE);
+    this._resizeTimeout = setTimeout(() => {
+      this._eventbus.broadcast(TimelineEventNames.RESIZE);
     }, 200);
   }
 
   private _importSystemEntryWithEventbusDependency(systemName: string): any {
     const ctor = this._importSystemEntry(systemName);
-    return new ctor(this.eventbus);
+    return new ctor(this._eventbus);
   }
 
   private _importSystemEntry(systemName: string): any {
-    return this.importer.import(systemName)[systemName];
+    return this._importer.import(systemName)[systemName];
   }
 
   private _requestInstanceHandler(
@@ -151,7 +190,7 @@ export class EngineFactory implements IEngineFactory {
     systemName: string,
     resultCallback: TResultCallback
   ) {
-    const action = this.actionsLookup[systemName];
+    const action = this._actionsLookup[systemName];
     if (action) {
       resultCallback(action);
     } else {
@@ -171,37 +210,41 @@ export class EngineFactory implements IEngineFactory {
       undefined;
     if (configuration.eventActions?.length) {
       actionRegistryListener = new ActionRegistryEventbusListener();
-      this.eventbus.registerEventlistener(actionRegistryListener);
+      this._eventRemovers.push(
+        this._eventbus.registerEventlistener(actionRegistryListener)
+      );
     }
 
-    this.eventbus.registerInterceptor(
-      TimelineEventNames.REQUEST_TIMELINE_URI,
-      new RequestVideoUriInterceptor(this.eventbus)
+    this._eventRemovers.push(
+      this._eventbus.registerInterceptor(
+        TimelineEventNames.REQUEST_TIMELINE_URI,
+        new RequestVideoUriInterceptor(this._eventbus)
+      )
     );
 
     resolver =
-      resolver || new ConfigurationResolver(this.importer, this.eventbus);
+      resolver || new ConfigurationResolver(this._importer, this._eventbus);
     const [actionLookup, resolvedConfiguration] = resolver.process(
       configuration,
       actionRegistryListener
     );
-    this.actionsLookup = actionLookup;
+    this._actionsLookup = actionLookup;
 
     const timelineProviders = this._createTimelineProviders(
       resolvedConfiguration,
-      this.eventbus
+      this._eventbus
     );
 
     const { language, labels } = configuration;
     const languageManager = new LanguageManager(
       language,
       labels,
-      this.eventbus
+      this._eventbus
     );
 
     const engineInstance = new EngineClass(
       resolvedConfiguration,
-      this.eventbus,
+      this._eventbus,
       timelineProviders,
       languageManager
     );